function-source/requirements.txt
requests
google-cloud
google-cloud-bigquery-datatransfer
protobuf
google-cloud-datastore
google-cloud-firestore
